# src/main/resources/api/openapi.yaml
openapi: 3.0.4
info:
  title: Sample Ecommerce App
  description: "'This is a ***sample ecommerce app API***.  You can find out more\
    \ about Swagger at [swagger.io](http://swagger.io). Description supports markdown\
    \ markup. For example, you can use the `inline code` using back ticks.'\n"
  termsOfService: https://github.com/chrisgya/webflux-r2dbc/blob/master/LICENSE
  contact:
    email: chrisgya500@gmail.com
  license:
    name: MIT
    url: https://github.com/chrisgya/webflux-r2dbc/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://ecommerce.swagger.io/v3
...
paths:
...
  /api/v1/orders:
    get:
      tags:
        - order
      summary: Returns the orders of given user
      description: Returns orders of given user
      operationId: getOrdersByCustomerId
      parameters:
        - name: customerId
          in: query
          description: Customer Identifier
          required: true
          style: form
          explode: true
          schema:
            type: string
      responses:
        "200":
          description: If order exists.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        "404":
          description: Order doesn't exist for given user.
          content: {}
    post:
      tags:
        - order
      summary: Creates a new order for the given order request
      description: Creates a new order for the given order request.
      operationId: addOrder
      requestBody:
        description: New Order Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
      responses:
        "201":
          description: Order added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "406":
          description: If payment is not authorized.
          content: {}
  /api/v1/orders/{id}:
    get:
      tags:
        - order
      summary: Returns the order of given order ID
      description: Returns orders of given order ID
      operationId: getByOrderId
      parameters:
        - name: id
          in: path
          description: Order Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: If order exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: Order doesn't exist for given user.
          content: {}
...
  /api/v1/products:
    get:
      tags:
        - product
      summary: Returns all the matched products
      description: Returns the products that matches the given query criteria
      operationId: queryProducts
      parameters:
        - name: tag
          in: query
          description: Product tag
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: name
          in: query
          description: Product name
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: page
          in: query
          description: Query page number
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          description: Query page size
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/v1/products/{id}:
    get:
      tags:
        - product
      summary: Returns a product
      description: Returns the product that matches the given product ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          description: Product Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: string
          description: Item Identifier
        quantity:
          type: integer
          description: The item quantity
          format: int32
        unitPrice:
          type: BigDecimal
          description: The item's price per unit
      description: Items in shopping cart
    Order:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        customer:
          $ref: '#/components/schemas/User'
        address:
          $ref: '#/components/schemas/Address'
        card:
          $ref: '#/components/schemas/Card'
        date:
          type: string
          description: Order's data and time details
          format: date-time
        items:
          type: array
          description: Collection of order items.
          items:
            $ref: '#/components/schemas/Item'
        total:
          type: BigDecimal
          description: Order total
        payment:
          $ref: '#/components/schemas/Payment'
        shipment:
          $ref: '#/components/schemas/Shipment'
        status:
          type: string
          description: Order Status
          enum:
            - CREATED
            - PAID
            - SHIPPED
            - PAYMENT_FAILED
            - SHIPMENT_FAILED
            - COMPLETED
      description: Represents an order
      xml:
        name: Order
...
    Address:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        number:
          type: string
          description: house of flat number
        residency:
          type: string
          description: Socieity or building name
        street:
          type: string
          description: street name
        city:
          type: string
          description: city name
        state:
          type: string
          description: state name
        country:
          type: string
          description: country name
        pincode:
          type: string
          description: postal code
      xml:
        name: Address
    Card:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        cardNumber:
          type: string
          description: Card Number
        expires:
          type: string
          description: Expiration date
        cvv:
          type: string
          description: CVV code
        userId:
          type: string
      xml:
        name: Card
...
    Payment:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        authorized:
          type: boolean
          description: Flag that specified whether payment is authorized or not
        message:
          type: string
          description: Approval or rejection message
      xml:
        name: Payment
    Shipment:
      type: object
      properties:
        id:
          type: string
          description: Shipment Identifier
        carrier:
          type: string
          description: Shipping Carrier
        trackingNumber:
          type: string
          description: Shipping Tracking Number
        estDeliveryDate:
          type: string
          description: Estimated Delivery Date
          format: date
      xml:
        name: Shipment
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: string
      xml:
        name: User
    NewOrder:
      type: object
      properties:
        customerId:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        card:
          $ref: '#/components/schemas/Card'
        items:
          type: array
          description: Items in shopping cart
          items:
            $ref: '#/components/schemas/Item'
      description: Contains the new order request information
      xml:
        name: NewOrder
...
    Product:
      type: object
      properties:
        id:
          type: string
          description: Product identifier
        name:
          type: string
          description: Product Name
        description:
          type: string
          description: Prodcut's description
        imageUrl:
          type: string
          description: Product image's URL
        price:
          type: BigDecimal
          description: Product price
        count:
          type: integer
          description: Product count
          format: int32
        tag:
          type: array
          description: Tags associated with the product
          items:
            $ref: '#/components/schemas/Tag'
      description: Product information
    Tag:
      type: object
      properties:
        id:
          type: string
          description: Tag identifier
        name:
          type: string
          description: Tag Name
      description: Tag information